requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
wincertstore==0.2
flask-marshmallow
validate-docbr==1.7.0
flask-mongoengine==0.9.5
simplejson